updated cart

+++ src/main/java/com/example/elearningwebm4/backend/controllers/CartController.java
@@ -6,12 +6,8 @@ import com.example.elearningwebm4.backend.models.Users;
 import com.example.elearningwebm4.backend.services.CartItemService;
 import com.example.elearningwebm4.backend.services.CartService;
 import com.example.elearningwebm4.backend.services.UsersService;
-import com.example.elearningwebm4.backend.services.serviceimpl.CustomUserDetailsService;
-import jakarta.servlet.http.HttpServletRequest;
-import org.apache.tomcat.util.net.openssl.ciphers.Authentication;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
@@ -35,17 +31,11 @@ public class CartController {
 
     @GetMapping("/cart-view")
     public String cartView(Model model, @AuthenticationPrincipal UserDetails userDetails) {
-//        Long userId = (Long) model.getAttribute("userId");
-//        if (userId == null) {
-//            return "redirect:/login";
-//        }
-        // HỎI ANH CÔNG LÀ DÙNG @AUTHEN VỚI TẠO 1 CONTROLLERGLOBAL THÌ CÁI NÀO OK HƠN
-        Users users = usersService.findByEmail(userDetails.getUsername());
         if (userDetails == null) {
             return "redirect:/login";
         }
-//        Cart cart = cartService.findByUserId(userId);
-        Cart cart = cartService.findByUserId(users.getUserId());
+        Users user = usersService.findByEmail(userDetails.getUsername());
+        Cart cart = cartService.findByUserId(user.getUserId());
         if (cart.getItems().isEmpty()) {
             return "cart/cart-view";
         }
@@ -62,38 +52,42 @@ public class CartController {
     @PostMapping("/add-to-cart")
     public String addToCart(@RequestParam Long courseId,
                             RedirectAttributes redirectAttributes,
-                            Model model) {
-        Long userId = (Long) model.getAttribute("userId");
-        if (userId == null) {
+                            @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
             return "redirect:/login";
         }
-        Cart cart = cartService.findByUserId(userId);
+        Users user = usersService.findByEmail(userDetails.getUsername());
+        Cart cart = cartService.findByUserId(user.getUserId());
         try {
             cartItemService.addCourseToCart(cart.getCartId(), courseId);
         } catch (Exception e) {
             redirectAttributes.addFlashAttribute("message", "Khóa học đã có trong giỏ hàng!");
             return "redirect:/";
         }
-
         return "redirect:/";
     }
 
     @PostMapping("/remove-cart-item")
-    public String removeCartItem(@RequestParam(name = "courseId") Long courseId,
-                                 Model model) {
-        Long userId = (Long) model.getAttribute("userId");
-        Cart cart = cartService.findByUserId(userId);
-        System.out.println(cart.getCartId());
+    public String removeCartItemFromCart(@RequestParam(name = "courseId") Long courseId,
+                                         @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return "redirect:/login";
+        }
+        Users user = usersService.findByEmail(userDetails.getUsername());
+        Cart cart = cartService.findByUserId(user.getUserId());
         cartItemService.removeCartItem(cart.getCartId(), courseId);
         return "redirect:/cart/cart-view";
     }
 
     @PostMapping("/payment")
     public String payCart(RedirectAttributes redirectAttributes,
-                          Model model) {
-        Long userId = (Long) model.getAttribute("userId");
-        cartService.payCart(userId);
-        redirectAttributes.addFlashAttribute("message","Thanh toán thành công giỏ hàng!");
+                          @AuthenticationPrincipal UserDetails userDetails) {
+        if (userDetails == null) {
+            return "redirect:/login";
+        }
+        Users user = usersService.findByEmail(userDetails.getUsername());
+        cartService.payCart(user.getUserId());
+        redirectAttributes.addFlashAttribute("message", "Thanh toán thành công giỏ hàng!");
         return "redirect:/";
     }
 }
